changed WalletService class and created interface
